Relocate App component to pages folder, update import statment in src/index.js
